CP-49690 Add missing sm-core-libs methods
For any methods which were new in sm-core-libs, recreate them back in
the sm-libs code.

Where necessary, adjust method signatures used within sm-libs to match
what the other packages are expecting, for those cases where
sm-core-libs altered the signatures of otherwise identical methods.
