+++ libs/sm/core/f_exceptions.py
@@ -1,2 +1,3 @@
 # Compatibility layer for code relying on the old sm-core-libs
-from sm.core.xs_errors import XenError
+from sm.core.xs_errors import SRException
+from sm.core.xs_errors import SROSError as XenError
